add: added authentication to movie routes to authorize

+++ src/routers/Movies.js
@@ -58,7 +58,7 @@ router.post('/upload', auth,  upload.any(), async function(req, res, next) {
  });
 
 /* Edit A Movie Entry*/
-router.post('/edit/:id', function(req, res, next) {
+router.post('/edit/:id', auth,  function(req, res, next) {
     data = {
       title: req.body.title,
       description: req.body.description,
@@ -66,7 +66,7 @@ router.post('/edit/:id', function(req, res, next) {
       timestamps: Date.now()
     }
 
-    Movie.findByIdAndUpdate({_id: req.params.id}, data, (error, movie) => {
+    Movie.findByIdAndUpdate({_id: req.params.id, owner: req.user._id}, data, (error, movie) => {
       if (error) {
           return res.send({
               success: false,
@@ -82,43 +82,56 @@ router.post('/edit/:id', function(req, res, next) {
 });
 
 /* Get All Movies */
-router.get('/', function(req, res, next) {
-    Movie.find({}, (error, movie) => {
-      if (error) {
-          return res.send({
-              success: false,
-              message: error
-          })
-      } else {
-          return res.send({
-              success: true,
-              message: movie
-          })
-      }
-  })
-  });
+router.get('/', auth,  async function(req, res, next) {
+  try{
+    await req.user.populate('movies').execPopulate()
+    res.status(200).json({
+      movies: req.user.movies,
+      success: true
+    })
+  }catch(err) {
+    res.status(400).json({
+      message: 'can not fetch movies created by user',
+      success: false
+    })
+  }
 
+})
+  
   /* Get A Movie */
-  router.get('/:id', function(req, res, next) {
-    var data = req.query
-    Movie.find({_id: req.params.id}, (error, movie) => {
-      if (error) {
-          return res.send({
-              success: false,
-              message: error
-          })
-      } else {
-          return res.send({
-              success: true,
-              message: movie
-          })
-      }
+router.get('/:id', async function(req, res) {
+  const _id = req.params._id
+
+  try{
+    const movie = await Movie.findOne({id})
+
+    if(!movie) {
+      res.status(404).json({
+        message: 'movie was found successfully',
+        success: false
+      })
+    }
+
+    res.status(200).json({
+      movie,
+      success: true
+    })
+  }catch(e){
+    res.status(400).json({
+      message: 'unable to link user with movie search',
+      success: false
+    })
+  }
+  
+}, (error, req, res, next)=> {
+  res.status.json({
+    message: error.message
   })
-  });
+});
 
   /* Delete Movie */
-  router.delete('/delete/:id', (req, res, next) => {
-    Movie.findOneAndDelete({_id: req.params.id}, (error, movie) => {
+  router.delete('/delete/:id', auth, (req, res, next) => {
+    Movie.findOneAndDelete({_id: req.params.id, owner: req.user._id}, (error, movie) => {
       if (error) {
         return res.send({
           success: false,
